restrict shop related tables from updating the code and delete if there are linked products

+++ app/Http/Controllers/Admin/BrandController.php
@@ -4,11 +4,11 @@ namespace App\Http\Controllers\Admin;
 
 use App\Helpers\CollectionHelper;
 use App\Helpers\ResponseHelper;
-use App\Helpers\StringHelper;
 use App\Http\Controllers\Controller;
+use App\Repositories\Shop\Brand\BrandCreateRequest;
 use App\Repositories\Shop\Brand\BrandRepositoryInterface;
+use App\Repositories\Shop\Brand\BrandUpdateRequest;
 use Illuminate\Database\QueryException;
-use Illuminate\Validation\Rule;
 
 class BrandController extends Controller
 {
@@ -31,20 +31,20 @@ class BrandController extends Controller
         return $this->brandRepository->find($slug);
     }
 
-    public function store()
+    public function store(BrandCreateRequest $request)
     {
         try {
-            $brand = $this->brandRepository->create(self::validateCreate());
+            $brand = $this->brandRepository->create($request->validated());
             return response()->json($brand, 201);
         } catch (QueryException $e) {
             return ResponseHelper::generateValidateError('code', 'The code has already been taken.');
         }
     }
 
-    public function update($slug)
+    public function update(BrandUpdateRequest $request, $slug)
     {
         try {
-            return $this->brandRepository->update($slug, self::validateUpdate($slug));
+            return $this->brandRepository->update($slug, $request->validated());
         } catch (QueryException $e) {
             return ResponseHelper::generateValidateError('code', 'The code has already been taken.');
         }
@@ -52,36 +52,10 @@ class BrandController extends Controller
 
     public function destroy($slug)
     {
-        return response()->json(['message' => 'Permission denied.'], 403);
+        if ($this->brandRepository->checkProducts($slug)) {
+            return response()->json(['message' => 'Cannot delete brand if there is a linked product.'], 403);
+        }
 
         return $this->brandRepository->delete($slug);
-    }
-
-    private static function validateCreate()
-    {
-        request()->merge(['slug' => StringHelper::generateUniqueSlug()]);
-
-        return request()->validate([
-            'code' => 'required|unique:brands|size:4',
-            'slug' => 'required|unique:brands',
-            'name' => 'required|unique:brands',
-            'image_slug' => 'nullable|exists:App\Models\File,slug',
-        ]);
-    }
-
-    private static function validateUpdate($slug)
-    {
-        return request()->validate([
-            'code' => [
-                'required',
-                // Rule::unique('brands')->ignore($slug, 'slug'),
-                'size:4',
-            ],
-            'name' => [
-                'required',
-                Rule::unique('brands')->ignore($slug, 'slug'),
-            ],
-            'image_slug' => 'nullable|exists:App\Models\File,slug',
-        ]);
     }
 }
