+++ app/Http/Controllers/Admin/ShopCategoryController.php
@@ -4,11 +4,11 @@ namespace App\Http\Controllers\Admin;
 
 use App\Helpers\CollectionHelper;
 use App\Helpers\ResponseHelper;
-use App\Helpers\StringHelper;
 use App\Http\Controllers\Controller;
+use App\Repositories\Shop\ShopCategory\ShopCategoryCreateRequest;
 use App\Repositories\Shop\ShopCategory\ShopCategoryRepositoryInterface;
+use App\Repositories\Shop\ShopCategory\ShopCategoryUpdateRequest;
 use Illuminate\Database\QueryException;
-use Illuminate\Validation\Rule;
 
 class ShopCategoryController extends Controller
 {
@@ -31,32 +31,20 @@ class ShopCategoryController extends Controller
         return $this->shopCategoryRepository->find($slug)->load(['shopMainCategory', 'shopSubCategories']);
     }
 
-    public function store()
+    public function store(ShopCategoryCreateRequest $request)
     {
-        $validatedData = self::validateCreate();
-
-        if (request('shop_main_category_slug')) {
-            $validatedData['shop_main_category_id'] = $this->shopCategoryRepository->getMainCategoryIdBySlug(request('shop_main_category_slug'));
-        }
-
         try {
-            $shopCategory = $this->shopCategoryRepository->create($validatedData)->refresh()->load(['shopMainCategory']);
+            $shopCategory = $this->shopCategoryRepository->create($request->validated())->refresh()->load(['shopMainCategory']);
             return response()->json($shopCategory, 201);
         } catch (QueryException $e) {
             return ResponseHelper::generateValidateError('code', 'The code has already been taken for this product type.');
         }
     }
 
-    public function update($slug)
+    public function update(ShopCategoryUpdateRequest $request, $slug)
     {
-        $validatedData = self::validateUpdate($slug);
-
-        if (request('shop_main_category_slug')) {
-            $validatedData['shop_main_category_id'] = $this->shopCategoryRepository->getMainCategoryIdBySlug(request('shop_main_category_slug'));
-        }
-
         try {
-            return $this->shopCategoryRepository->update($slug, $validatedData);
+            return $this->shopCategoryRepository->update($slug, $request->validated());
         } catch (QueryException $e) {
             return ResponseHelper::generateValidateError('code', 'The code has already been taken for this product type.');
         }
@@ -64,7 +52,9 @@ class ShopCategoryController extends Controller
 
     public function destroy($slug)
     {
-        return response()->json(['message' => 'Permission denied.'], 403);
+        if ($this->shopCategoryRepository->checkProducts($slug)) {
+            return response()->json(['message' => 'Cannot delete category if there is a linked product.'], 403);
+        }
 
         return $this->shopCategoryRepository->delete($slug);
     }
@@ -90,31 +80,5 @@ class ShopCategoryController extends Controller
         return $this->shopCategoryRepository->update($slug, request()->validate([
             'search_index' => 'required|numeric',
         ]));
-    }
-
-    private static function validateCreate()
-    {
-        request()->merge(['slug' => StringHelper::generateUniqueSlug()]);
-
-        return request()->validate([
-            'code' => 'required|size:3',
-            'slug' => 'required|unique:shop_categories',
-            'name' => 'required|unique:shop_categories',
-            'image_slug' => 'nullable|exists:App\Models\File,slug',
-            'shop_main_category_slug' => 'nullable|exists:App\Models\ShopMainCategory,slug',
-        ]);
-    }
-
-    private static function validateUpdate($slug)
-    {
-        return request()->validate([
-            'code' => 'required|size:3',
-            'name' => [
-                'required',
-                Rule::unique('shop_categories')->ignore($slug, 'slug'),
-            ],
-            'image_slug' => 'nullable|exists:App\Models\File,slug',
-            'shop_main_category_slug' => 'nullable|exists:App\Models\ShopMainCategory,slug',
-        ]);
     }
 }
