+++ app/Http/Controllers/Admin/ShopSubCategoryController.php
@@ -4,11 +4,11 @@ namespace App\Http\Controllers\Admin;
 
 use App\Helpers\CollectionHelper;
 use App\Helpers\ResponseHelper;
-use App\Helpers\StringHelper;
 use App\Http\Controllers\Controller;
+use App\Repositories\Shop\ShopSubCategory\ShopSubCategoryCreateRequest;
 use App\Repositories\Shop\ShopSubCategory\ShopSubCategoryRepositoryInterface;
+use App\Repositories\Shop\ShopSubCategory\ShopSubCategoryUpdateRequest;
 use Illuminate\Database\QueryException;
-use Illuminate\Validation\Rule;
 
 class ShopSubCategoryController extends Controller
 {
@@ -31,43 +31,30 @@ class ShopSubCategoryController extends Controller
         return $this->subCategoryRepository->find($slug)->load(['shopCategory']);
     }
 
-    public function store()
+    public function store(ShopSubCategoryCreateRequest $request)
     {
-        $validatedData = self::validateCreate();
-        $validatedData['shop_category_id'] = $this->subCategoryRepository->getShopCategoryIdBySlug(request('shop_category_slug'));
-
         try {
-            $shopCategory = $this->subCategoryRepository->create($validatedData)->refresh()->load(['shopCategory']);
+            $shopCategory = $this->subCategoryRepository->create($request->validated())->refresh()->load(['shopCategory']);
             return response()->json($shopCategory, 201);
         } catch (QueryException $e) {
             return ResponseHelper::generateValidateError('code', 'The code has already been taken for this category.');
         }
     }
 
-    public function update($slug)
+    public function update(ShopSubCategoryUpdateRequest $request, $slug)
     {
-        $validatedData = self::validateUpdate($slug);
-        $validatedData['shop_category_id'] = $this->subCategoryRepository->getShopCategoryIdBySlug(request('shop_category_slug'));
-
         try {
-            $subCategory = $this->subCategoryRepository->update($slug, $validatedData);
+            return $this->subCategoryRepository->update($slug, $request->validated())->load(['shopCategory']);
         } catch (QueryException $e) {
             return ResponseHelper::generateValidateError('code', 'The code has already been taken for this category.');
         }
-
-        // Update the category ids of related products
-        foreach ($subCategory->products as $product) {
-            $product->update([
-                'shop_category_id' => $validatedData['shop_category_id'],
-            ]);
-        }
-
-        return $subCategory->load(['shopCategory'])->unsetRelation('products');
     }
 
     public function destroy($slug)
     {
-        return response()->json(['message' => 'Permission denied.'], 403);
+        if ($this->subCategoryRepository->checkProducts($slug)) {
+            return response()->json(['message' => 'Cannot delete sub category if there is a linked product.'], 403);
+        }
 
         return $this->subCategoryRepository->delete($slug);
     }
@@ -96,29 +83,5 @@ class ShopSubCategoryController extends Controller
         return $this->subCategoryRepository->update($slug, request()->validate([
             'search_index' => 'required|numeric',
         ]));
-    }
-
-    private static function validateCreate()
-    {
-        request()->merge(['slug' => StringHelper::generateUniqueSlug()]);
-
-        return request()->validate([
-            'code' => 'required|size:2',
-            'slug' => 'required|unique:shop_sub_categories',
-            'name' => 'required|unique:shop_sub_categories',
-            'shop_category_slug' => 'required|exists:App\Models\ShopCategory,slug',
-        ]);
-    }
-
-    private static function validateUpdate($slug)
-    {
-        return request()->validate([
-            'code' => 'required|size:2',
-            'name' => [
-                'required',
-                Rule::unique('shop_sub_categories')->ignore($slug, 'slug'),
-            ],
-            'shop_category_slug' => 'required|exists:App\Models\ShopCategory,slug',
-        ]);
     }
 }
